+++ torchtitan/components/quantization/mx.py
@@ -10,11 +10,13 @@ from importlib.util import find_spec
 from typing import Any, ClassVar, Literal
 
 import torch.nn as nn
-from torchtitan.components.quantization import MXFP8_GROUP_ALIGNMENT_SIZE
+from torchtitan.components.quantization import (
+    MXFP8_GROUP_ALIGNMENT_SIZE,
+    QuantizationConverter,
+)
 
 from torchtitan.distributed import ParallelDims
 from torchtitan.models.common.moe.utils import set_token_group_alignment_size_m
-from torchtitan.protocols.model_converter import QuantizationConverter
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability, has_rocm_capability
